web/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
websockets==12.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
numpy==1.24.3
pillow==10.0.0
python-magic==0.4.27

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
psutil==5.9.6
pydantic-settings==2.1.0